rename solar localization
